% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateLogitShare.R
\name{aggregateLogitShare}
\alias{aggregateLogitShare}
\title{Aggregate the logit shares across hs}
\usage{
aggregateLogitShare(logitShare, renovationIn)
}
\arguments{
\item{name}{description}
}
\description{
Aggregate the logit shares across hs
}
\author{
Ricarda Rosemann
}
