% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeBrickShare.R
\name{computeBrickShare}
\alias{computeBrickShare}
\title{Estimate Brick model heating system shares}
\usage{
computeBrickShare(variable, data)
}
\arguments{
\item{variable}{character, Brick variable that is evaluated}

\item{data}{data frame, Brick stock and flow data}
}
\description{
Estimate Brick model heating system shares
}
\author{
Ricarda Rosemann
}
