% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDiscountSum.R
\name{computeDiscountSum}
\alias{computeDiscountSum}
\title{Compute the discounted sum of a given variable between two time periods}
\usage{
computeDiscountSum(df, dfDt, dfDiscount, ttotNum, ttot2Num)
}
\arguments{
\item{df}{data frame, should contain one time vector}

\item{dfDt}{data frame, contains lengths of time periods}

\item{dfDiscount}{data frame, contains discount rate}

\item{ttot}{numeric, first time vector}

\item{ttot2}{numeric, second time vector}
}
\description{
Compute the discount rate between two time periods
and determine the discounted sum of a given variable.
This can be used to compute the expected value of a time dependent variable,
where the end time is randomly distributed (e.g. for life cycle operational costs)
}
\author{
Ricarda Rosemann
}
