% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLCC.R
\name{computeLCC}
\alias{computeLCC}
\title{Compute LCC}
\usage{
computeLCC(dfLt, dfCostsOpe, dfCostsRen, dfDt, dfDiscount)
}
\arguments{
\item{dfLt}{data frame, lifetime estimate}

\item{dfCostsOpe}{data frame, operational costs}

\item{dfCostsRen}{data frame, installation/renovation costs}

\item{dfDt}{data frame, lengths of time periods}

\item{dfDiscount}{data frame, discount factors}
}
\description{
Compute the lifecycle costs (LCC)
}
\author{
Ricarda Rosemann
}
