% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLCCVaryDiscount.R
\name{computeLCCVaryDiscount}
\alias{computeLCCVaryDiscount}
\title{Compute LCC for varying discount rates}
\usage{
computeLCCVaryDiscount(dfLt, dfCostsOpe, dfCostsRen, dfDt, dfDiscount)
}
\arguments{
\item{dfLt}{data frame, lifetime estimate}

\item{dfCostsOpe}{data frame, operational costs}

\item{dfCostsRen}{data frame, installation/renovation costs}

\item{dfDt}{data frame, lengths of time periods}

\item{dfDiscount}{data frame, discount factors}
}
\description{
This was a function to test the effect of a varying discount rate and perspectively
can be removed
}
\author{
Ricarda Rosemann
}
