% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLCOH.R
\name{computeLCOH}
\alias{computeLCOH}
\title{Compute LCOH from LCC}
\usage{
computeLCOH(dfLcc, dfLt, dfUe, dfDt, dfDiscount, dims)
}
\arguments{
\item{dfLcc}{data frame, lifecycle costs}

\item{dfLt}{data frame, lifetime estimate}

\item{dfUe}{data frame, useful energy demand}

\item{dfDt}{data frame, lengths of time periods}

\item{dfDiscount}{data frame, discount rates}

\item{dims}{character, dimensions of all data, excluding time dimension}
}
\description{
Compute LCOH from LCC
}
\author{
Ricarda Rosemann
}
