% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLogitShare.R
\name{computeLogitShare}
\alias{computeLogitShare}
\title{Compute heating system shares from logit model}
\usage{
computeLogitShare(variable, lcc, lambda)
}
\arguments{
\item{variable}{character, flow variable for which the share is computed}

\item{lcc}{data frame, lifecycle costs}

\item{lambda}{numeric, price sensitivity}
}
\description{
Based on the lifecycle costs, compute the shares we would expect for each heating system
if this was simply a logit model
}
\author{
Ricarda Rosemann
}
