% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLt.R
\name{computeLt}
\alias{computeLt}
\title{Compute estimated lifetimes of heating systems}
\usage{
computeLt(ltMax, outflow, brickRes, conShare, ttotNum)
}
\arguments{
\item{ltMax}{data frame, maximum lifetimes computed ex-ante from Weibull distribution}

\item{outflow}{data frame, brick outflows}

\item{brickRes}{named list of data frames, brick output data on stocks and flows}

\item{conShare}{data frame, share of constructions in total inflows}

\item{ttotNum}{numeric, all time periods}
}
\description{
Determine heating system lifetimes such that they roughly follow a Weibull distribution,
but match Brick outputs.
}
\author{
Ricarda Rosemann
}
