% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtMixed.R
\name{computeLtMixed}
\alias{computeLtMixed}
\title{Compute mixed lifetimes of heating systems}
\usage{
computeLtMixed(ltAnte, outflow, brickRes, conShare, ttotNum, dims)
}
\arguments{
\item{ltAnte}{data frame, ex-ante lifetimes}

\item{outflow}{data frame, brick outflows}

\item{brickRes}{named list of data frames, brick output data on stocks and flows}

\item{conShare}{data frame, share of constructions in total inflows}

\item{ttotNum}{numeric, all time periods}

\item{dims}{character, dimensions of the data without time periods}
}
\description{
Determine heating system lifetimes such that they roughly follow a Weibull distribution,
but match Brick outputs.
}
\author{
Ricarda Rosemann
}
