% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtPost.R
\name{computeLtPost}
\alias{computeLtPost}
\title{Compute ex-post lifetimes}
\usage{
computeLtPost(inflow, outflow, data, conShare, p_ttotVin, ttotNum, dims)
}
\arguments{
\item{inflow}{data frame, total inflows}

\item{outflow}{data frame, total outflows}

\item{data}{named list, stock and flow data from brick results}

\item{conShare}{data frame, share of constructions in total inflows}

\item{p_ttotVin}{data frame, mapping between time periods and vintages}

\item{ttotNum}{numeric, all time periods}

\item{dims}{character, dimensions of the data without time periods}
}
\description{
Compute ex-post lifetimes
}
\author{
Ricarda Rosemann
}
