% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convGDX2MIF.R
\name{convGDX2MIF}
\alias{convGDX2MIF}
\title{Read in GDX from BRICK and write *.mif reporting}
\usage{
convGDX2MIF(
  gdx,
  tmpl = NULL,
  file = NULL,
  scenario = "default",
  t = NULL,
  silent = TRUE
)
}
\arguments{
\item{gdx}{file path to a BRICK gdx}

\item{tmpl}{character, BRICK reporting template. There has to be a brickSets
mapping named with the same suffix: \code{brickSets_<tmpl>.yaml}}

\item{file}{name of the mif file which will be written, if no name is
provided a magpie object containing all the reporting information is
returned}

\item{scenario}{scenario name that is used in the *.mif reporting}

\item{t}{numeric vector of reporting periods (years)}

\item{silent}{boolean, suppress warnings and printing of dimension mapping}
}
\description{
Read in all information from GDX file that was generated with BRICK and
create the *.mif reporting
}
\author{
Robin Hasse
}
