% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEnergy.R
\name{.addCarrierDimension}
\alias{.addCarrierDimension}
\title{Add carrier dimesion based on heating system technology}
\usage{
.addCarrierDimension(v_stock, hsCarrier)
}
\arguments{
\item{v_stock}{MagPIE object, BRICK variable}

\item{hsCarrier}{data.frame, mapping between heating technology and energy
carrier}
}
\value{
MagPIE object with additional carrier dimension
}
\description{
Add carrier dimesion based on heating system technology
}
