% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.addDim}
\alias{.addDim}
\title{Add dimension of not yet existing}
\usage{
.addDim(x, dimName, item, dim = 3.1)
}
\arguments{
\item{x}{MAgPIE object which should be extended.}

\item{dimName}{The name of the new dimension}

\item{item}{One or more names of items in the new dimension.}

\item{dim}{The dimension number of the new dimension (e.g. 3.1)}
}
\value{
The extended MAgPIE object
}
\description{
wrapper around \code{magclass::addDim} that does nothing if a dimension of
the given name already exists
}
\author{
Robin Hasse
}
