% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLt.R
\name{.addResults}
\alias{.addResults}
\title{Add the final estimate results of a specific in and a specific out time period
to the data frame storing all final estimates}
\usage{
.addResults(dfOrig, dfNew, tIn, tOut)
}
\arguments{
\item{dfOrig}{data frame, contains all previous final estimates}

\item{dfNew}{data frame, newly computed final estimates}

\item{tIn}{numeric, time period of installation / inflow of the newly computed mixed estimates}

\item{tOut}{numeric, time period of removal / outflow of the newly computed mixed estimates}
}
\description{
Add the final estimate results of a specific in and a specific out time period
to the data frame storing all final estimates
}
