% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtMixed.R
\name{.addResultsMixed}
\alias{.addResultsMixed}
\title{Add the mixed estimate results of a specific in and a specific out time period
to the data frame storing all mixed estimates}
\usage{
.addResultsMixed(dfOrig, dfNew, tIn, tOut, dims, addDims = NULL)
}
\arguments{
\item{dfOrig}{data frame, contains all previous mixed estimates}

\item{dfNew}{data frame, newly computed mixed estimates}

\item{tIn}{numeric, time period of installation / inflow of the newly computed mixed estimates}

\item{tOut}{numeric, time period of removal / outflow of the newly computed mixed estimates}

\item{dims}{character, dimensions of the data without time periods}

\item{addDims}{character, additional dimensions present in the data}
}
\description{
Add the mixed estimate results of a specific in and a specific out time period
to the data frame storing all mixed estimates
}
