% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLt.R
\name{.adjustLtDirect}
\alias{.adjustLtDirect}
\title{Rename direct lifetime estimate and add zero columns}
\usage{
.adjustLtDirect(dfLtDirect)
}
\arguments{
\item{dfLtDirect}{data frame, direct lifetime estimates}
}
\description{
Zero columns are required to be included in summations later
}
