% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAgg.R
\name{.agg}
\alias{.agg}
\title{Aggregate}
\usage{
.agg(x, agg, silent = TRUE)
}
\arguments{
\item{x}{MagPIE object, BRICK object}

\item{agg}{named vector of dimensions to aggregate.}

\item{silent}{logical, suppress warnings and printing of dimension mapping}
}
\value{
aggregated MagPIE objects without sub dimensions in dim 3
}
\description{
Aggregate
}
