% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportHeatingSystemDetails.R
\name{.avgAlongDim}
\alias{.avgAlongDim}
\title{Compute data average along given dimensions}
\usage{
.avgAlongDim(df, weight, dimToAvg, joinCols, ...)
}
\arguments{
\item{df}{data frame with the data to be manipulated}

\item{weight}{data frame with weights to apply to the average}

\item{dimToAvg}{character column name of the column to average over}

\item{joinCols}{character, columns to join \code{df} and \code{weight} by}

\item{...}{named character vector with tuples to be renamed}
}
\description{
Compute data average along given dimensions
}
