% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMatching.R
\name{.checkForDeviation}
\alias{.checkForDeviation}
\title{Check if the tolerance is violated anywhere}
\usage{
.checkForDeviation(stockBal, name, tol = 0.01)
}
\arguments{
\item{stockBal}{data frame with evaluation of a stock balance equation}

\item{name}{character, name/identifier of the stock balance equation}

\item{tol}{numeric, tolerance to check the equation against}
}
\description{
Check if the tolerance is violated anywhere
}
