% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLCCShare.R
\name{.checkLifetimeResults}
\alias{.checkLifetimeResults}
\title{Check plausibility of lifetime results}
\usage{
.checkLifetimeResults(dfLt, dims)
}
\arguments{
\item{dfLt}{data frame with lifetime data}

\item{dims}{character, dimensions to group by}
}
\description{
Check plausibility of lifetime results
}
