% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLCCShare.R
\name{.cleanReadGdx}
\alias{.cleanReadGdx}
\title{Read data from gdx and clean}
\usage{
.cleanReadGdx(
  gdx,
  symbol,
  vinExists = NULL,
  renAllowed = NULL,
  asMagpie = FALSE
)
}
\arguments{
\item{vinExists}{data frame with existing vintage and time period combinations}

\item{renAllowed}{data frame with all possible renovation transitions}

\item{df}{data frame with data as read from a gdx}
}
\description{
Read data from gdx and clean
}
