% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAgg.R
\name{.collapseName}
\alias{.collapseName}
\title{collapse a character vector to variable name}
\usage{
.collapseName(name)
}
\arguments{
\item{name}{character vector}
}
\value{
collapsed name
}
\description{
Elements of the vector are collapsed with \code{|}. If the last element is
unit embraced in brackets, it is separated with a space.
}
