% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.combinations}
\alias{.combinations}
\title{All Combinations of dimension elements}
\usage{
.combinations(lst)
}
\arguments{
\item{lst}{names list of dimension entries}
}
\value{
character vector with all combinations of the dimension elements
  each separated by \code{.}
}
\description{
All Combinations of dimension elements
}
\author{
Robin Hasse
}
