% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.computeDeviation}
\alias{.computeDeviation}
\title{Compute the deviation to target data}
\usage{
.computeDeviation(df, dfTarget)
}
\arguments{
\item{df}{data frame, containing data from calibration run}

\item{dfTarget}{data frame, containing historical data as calibration target}
}
\value{
data frame where the value column contains the deviation from target data
}
\description{
Compute the deviation to target data
}
