% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.computeFlowSum}
\alias{.computeFlowSum}
\title{Compute the sum of construction and renovation flow values}
\usage{
.computeFlowSum(flows)
}
\arguments{
\item{flows}{named list of data frames, each containing flow quantities.
Entries need to include \code{construction}
and one entry with \code{renovation} as part of the name.}
}
\value{
data frame with the sum in the value column
}
\description{
Compute the sum of construction and renovation flow values
}
