% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtPost.R
\name{.computeLeaveInFlow}
\alias{.computeLeaveInFlow}
\title{Determine leave time of inflows}
\usage{
.computeLeaveInFlow(tIn, tOut, dfInAll, dfStock, dfIn, dfOut, dims)
}
\arguments{
\item{tIn}{numeric, time period of installation / inflow}

\item{tOut}{numeric, time period of removal / outflow}

\item{dfInAll}{data frame, contains ex-post estimates of all previous time periods}

\item{dfStock}{data frame, ex-post lifetime estimates of the initial stock}

\item{dfIn}{data frame, total inflows}

\item{dfOut}{data frame, total outflows}

\item{dims}{character, dimensions of the data without time periods}
}
\description{
Determine leave time of inflows
}
