% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtPost.R
\name{.computeLeaveInitStock}
\alias{.computeLeaveInitStock}
\title{Determine leave time of initial stock}
\usage{
.computeLeaveInitStock(tOut, dfStockInit, dfStock, dfOut, dims)
}
\arguments{
\item{tOut}{numeric, time period of removal / outflow}

\item{dfStockInit}{data frame, brick result data on initial stock}

\item{dfStock}{data frame, contains ex-post life time estimates of previous time periods}

\item{dfOut}{data frame, total outflow}

\item{dims}{character, dimensions of the data without time periods}
}
\description{
Determine leave time of initial stock
}
