% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtAnte.R
\name{.computeLtAnte}
\alias{.computeLtAnte}
\title{Compute the ex-ante lifetime of flows by numerical integration}
\usage{
.computeLtAnte(dfLt, n = 4, selfImplementation = FALSE)
}
\arguments{
\item{dfLt}{data frame, needs to contain input dimensions, each with intervals
for all time periods and with the Weibull paramters}

\item{n}{numeric, number of steps in the numerical integration}
}
\description{
Compute the ex-ante lifetime of flows by numerical integration
}
