% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtAnte.R
\name{.computeLtAnteSimple}
\alias{.computeLtAnteSimple}
\title{Compute the ex-ante lifetime by a simple approximation formula}
\usage{
.computeLtAnteSimple(dfLt, dims, t0, standingLifetime = 0, cutOffShare = 0.95)
}
\arguments{
\item{dfLt}{data frame, needs to contain input dimensions, each with intervals
for all time periods and with the Weibull paramters}

\item{dims}{character, dimensions of the data, without time steps}

\item{t0}{numeric, initial time period}

\item{standingLifetime}{numeric, assumed prior standing lifetime}

\item{cutOffShare}{numeric, share at which the distribution function jumps to 1}
}
\description{
Compute the ex-ante lifetime by a simple approximation formula
}
