% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLt.R
\name{.computeLtDirect}
\alias{.computeLtDirect}
\title{Compute direct lifetime estimate}
\usage{
.computeLtDirect(dfLtMax, outMax, outflow)
}
\arguments{
\item{dfLtMax}{data frame, ex-ante lifetimes}

\item{outMax}{data frame, total ex-ante outflows}

\item{outflow}{data frame, total actual brick outflows}
}
\description{
Compute the lifetime as the maximum estimate scaled by the ratio of maximum
and actual total outflows
}
