% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLt.R
\name{.computeLtFinal}
\alias{.computeLtFinal}
\title{Compute the final lifetime estimate}
\usage{
.computeLtFinal(
  dfLt,
  unattrOut,
  unattrTot,
  tIn,
  tOut,
  dfStock = NULL,
  dfShare = NULL
)
}
\arguments{
\item{dfLt}{data frame, final lifetime estimates of previous time steps,
direct estimates and remaining inflows of the current time step}

\item{unattrOut}{data frame with outflow that has not been attributed to any inflow}

\item{unattrTot}{data frame, total unattributed outflow}

\item{tIn}{numeric, current time period of installation / inflow}

\item{tOut}{numeric, current time period of removal / outflow}

\item{dfStock}{data frame, mixed estimates and additional attribution of the initial stock
(only required for flow computations)}

\item{dfShare}{data frame, share of the current flow in total inflows
(only required for flow computations)}
}
\description{
The final estimate is computed as the sum of the direct estimate and the total
unattributed outflow of the current time period, unless this surpasses the
remaining inflow. In the latter case the final lifetime estimate is given by
the total remaining inflow.
}
