% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtMixed.R
\name{.computeLtMixed}
\alias{.computeLtMixed}
\title{Compute the mixed lifetime estimate}
\usage{
.computeLtMixed(
  dfLtMixed,
  unattrTot,
  tIn,
  tOut,
  dims,
  dfStock = NULL,
  dfShare = NULL
)
}
\arguments{
\item{dfLtMixed}{data frame, mixed lifetime estimates of previous time steps,
direct estimates and remaining inflows of the current time step}

\item{unattrTot}{data frame, total unattributed outflow}

\item{tIn}{numeric, current time period of installation / inflow}

\item{tOut}{numeric, current time period of removal / outflow}

\item{dims, }{character, dimensions of the data without time periods}

\item{dfStock}{data frame, mixed estimates and additional attribution of the initial stock
(only required for flow computations)}

\item{dfShare}{data frame, share of the current flow in total inflows
(only required for flow computations)}
}
\description{
The mixed estimate is computed as the sum of the direct estimate and the total
unattributed outflow of the current time period, unless this surpasses the
remaining inflow. In the latter case the mixed lifetime estimate is given by
the total remaining inflow.
}
