% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtAnte.R
\name{.computeLtStockAnte}
\alias{.computeLtStockAnte}
\title{Compute the ex-ante lifetime of the initial stock}
\usage{
.computeLtStockAnte(dfLt, standingLifetime = 6)
}
\arguments{
\item{dfLt}{data frame, needs to contain input dimensions, each with intervals
for all time periods and with the Weibull paramters}

\item{standingLifetime}{numeric, assumed standing lifetime of the initial stock}
}
\description{
Compute the ex-ante lifetime of the initial stock
}
