% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.computeRatioSq}
\alias{.computeRatioSq}
\title{Compute the ratio of the squares for two data sets}
\usage{
.computeRatioSq(dfx, dfy, addSign = TRUE)
}
\arguments{
\item{dfx}{data frame containing the data of the nominator}

\item{dfy}{data frame containing the data of the denominator}

\item{addSign}{logical indicating whether a heuristic sign of the squares should be computed}
}
\value{
data frame with squared ratio
}
\description{
Compute the ratio of the squares for two data sets
}
