% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLt.R
\name{.computeRelativeLt}
\alias{.computeRelativeLt}
\title{Compute the relative value of the lifetime estimates}
\usage{
.computeRelativeLt(dfLt, dfBrickIn, dfMax)
}
\arguments{
\item{dfLt}{data frame, mixed lifetime estimates}

\item{dfBrickIn}{data frame, total inflows as given by brick results}

\item{dfMax}{data frame, ex-ante lifetime estimates}
}
\description{
By setting the final estimates in relation to the total inflows, compute the
probabilities of removal in a given time period.
If both the total inflow and the absolute lifetime estimate are close to zero,
set this to the maximum value.
}
