% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.computeRelDev}
\alias{.computeRelDev}
\title{Compute the relative deviation to calibration target data as the relative euclidean distance}
\usage{
.computeRelDev(dfDev, dfTarget, tCalib, notInTargetGrp = NULL)
}
\arguments{
\item{dfDev}{data frame, containing deviation between calibration result and target data}

\item{dfTarget}{data frame or a list of one or two data frames, containing target data.
The option to pass two data frames is used to combine construction and renovation flows.
If two data frames are passed, the first one is assumed to be construction, the second renovation}

\item{tCalib}{numerical/factor, calibration time periods to filter calibration target data}

\item{notInTargetGrp}{character vector, pass columns that target data should not be grouped by
when computing the sum of the squares, although they exist in the data}
}
\value{
data frame with relative deviation in value column
}
\description{
Compute the relative deviation to calibration target data as the relative euclidean distance
}
