% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.computeSumSq}
\alias{.computeSumSq}
\title{Compute the square root of the sum of the squares in a data frame}
\usage{
.computeSumSq(df, tCalib, rprt = "", addSign = TRUE)
}
\arguments{
\item{df}{data frame, containing the data to be evaluated}

\item{tCalib}{numeric, time periods the model was calibrated on}

\item{rprt}{character, column names for which the sum of the squares should be reported separately}

\item{addSign}{logical indicating whether the heuristic sign of the square should be computed}
}
\value{
data frame with the square root of the sum of squares as value column
}
\description{
Compute the square root of the sum of the squares in a data frame
}
