% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLt.R
\name{.computeUnattrOut}
\alias{.computeUnattrOut}
\title{Compute the unattributed outflow}
\usage{
.computeUnattrOut(dfLt, dfBrickIn, tOut)
}
\arguments{
\item{dfLt}{data frame, direct lifetime estimates of current time period and
final lifetime estimates of previous time periods.
Required (value) columns: \code{directVal}, \code{addAttr}, \code{value}.}

\item{dfBrickIn}{data frame, inflows in brick results}

\item{tOut}{numeric, time period of removal for which we are computing the
unattributed outflows}
}
\description{
Compute the remaining inflow that has not yet been attributed to any outflows
in the time periods before \code{tOut}.
If the direct estimate surpasses this amount, then the difference is the un-
attributed outflow.
}
