% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAgg.R
\name{.constructDimMapping}
\alias{.constructDimMapping}
\title{Construct dimension mapping (either aggregation or reporting) with explicit
set elements for each dimension}
\usage{
.constructDimMapping(agg, rprt, brickSets, silent)
}
\arguments{
\item{agg}{named vector of dimensions to aggregate}

\item{rprt}{named vector of dimensions to report individually}

\item{brickSets}{named list with definition of common set elements}

\item{silent}{boolean, suppress warnings and printing of dimension mapping}
}
\value{
nested named list with dimension mapping
}
\description{
Construct dimension mapping (either aggregation or reporting) with explicit
set elements for each dimension
}
