% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.embrace}
\alias{.embrace}
\title{Escape tag in curly brackets}
\usage{
.embrace(tag)
}
\arguments{
\item{tag}{character tag}
}
\value{
character, tag in curly brackets
}
\description{
Escape tag in curly brackets
}
\author{
Robin Hasse
}
