% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.excludeRows}
\alias{.excludeRows}
\title{Exclude data matching giving filter criteria from a data set}
\usage{
.excludeRows(data, exclude)
}
\arguments{
\item{data}{data frame with the data set}

\item{exclude}{named list with entries to be excluded from the data.
The name gives the column name from which the entries given by the value should
be removed.}
}
\value{
the given data set without the rows identified by the filter criteria
}
\description{
Exclude data matching giving filter criteria from a data set
}
