% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.expandDims}
\alias{.expandDims}
\title{Extend dimensions of a data frame by adding NA entries, add variable name}
\usage{
.expandDims(df, varName, allSets)
}
\arguments{
\item{df}{data frame to be extended}

\item{varName}{character, variable name to be added}

\item{allSets}{character, sets that need to be included as column names}
}
\value{
data frame
}
\description{
Extend dimensions of a data frame by adding NA entries, add variable name
}
