% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtAnte.R
\name{.extractInitTime}
\alias{.extractInitTime}
\title{Determine the starting year of each time period}
\usage{
.extractInitTime(ttot, ttotNum = NULL)
}
\arguments{
\item{ttot}{numeric, time periods}

\item{ttotNum}{numeric, additional vector of time periods that needs to start
earlier than \code{ttot}}
}
\description{
Determine the starting year of each time period
}
