% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportHeatingSystemDetails.R
\name{.filterAndUnselect}
\alias{.filterAndUnselect}
\title{Filter a column for a specific value and remove that column}
\usage{
.filterAndUnselect(df, col, val)
}
\arguments{
\item{df}{data frame with the data to be filtered}

\item{col}{character, name of column to be filtered}

\item{val}{single value of varying type, value to filter \code{col} for}
}
\description{
Filter a column for a specific value and remove that column
}
