% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convGDX2MIF.R
\name{.findInconsistenSetElements}
\alias{.findInconsistenSetElements}
\title{Find inconsistencies in set elements between reporting template and gdx}
\usage{
.findInconsistenSetElements(brickSets, gdx)
}
\arguments{
\item{brickSets}{character, BRICK reporting template}

\item{gdx}{file path to a BRICK gdx}
}
\value{
named list of missing and surplus sets elements
}
\description{
Find inconsistencies in set elements between reporting template and gdx
}
