% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAgg.R
\name{.findTags}
\alias{.findTags}
\title{Find all tags in name escaped in curly brackets}
\usage{
.findTags(name)
}
\arguments{
\item{name}{character, variable name}
}
\value{
vector of tags in name, NULL if there are none
}
\description{
Find all tags in name escaped in curly brackets
}
