% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.getDims}
\alias{.getDims}
\title{Get dimension names from stock and flow objects}
\usage{
.getDims(calibObj, removeVin)
}
\arguments{
\item{calibObj}{list, calibration targets}

\item{removeVin}{logical, whether vintage dimension should be removed}
}
\description{
Gives the column names of a list of data frames except the first and the last
one, i.e. \code{qty} and \code{value}.
}
