% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAgg.R
\name{.missingElements}
\alias{.missingElements}
\title{List any missing elements in any given dimension}
\usage{
.missingElements(x, dimLst)
}
\arguments{
\item{x}{MagPIE object, BRICK object}

\item{dimLst}{named vector, dimension ans names and elements as values}
}
\value{
vector of missing dimension elements in x
}
\description{
List any missing elements in any given dimension
}
