% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtMixed.R
\name{.prepareLtAnte}
\alias{.prepareLtAnte}
\title{Prepare ex-ante lifetime data to be used in mixed lifetime computation}
\usage{
.prepareLtAnte(df, toRename, sumTtot = FALSE)
}
\arguments{
\item{df}{data frame, lifetime data to manipulate}

\item{toRename}{named character, old (value) and new name (name) of columns to be renamed}

\item{sumTtot}{logical, whether to sum over ttot}
}
\description{
Remove not required columns, rename columns and sum over ttot as indicated
}
