% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLt.R
\name{.prepareLtMax}
\alias{.prepareLtMax}
\title{Prepare maximum lifetime data to be used in estimated lifetime computation}
\usage{
.prepareLtMax(df, sumTtot = FALSE)
}
\arguments{
\item{df}{data frame, lifetime data to manipulate}

\item{sumTtot}{logical, whether to sum over ttotIn}
}
\description{
Remove not required columns, rename columns and sum over ttotIn as indicated
}
