% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtMixed.R
\name{.prepareLtMixed}
\alias{.prepareLtMixed}
\title{Rename direct lifetime estimate and add zero columns}
\usage{
.prepareLtMixed(dfLtDirect)
}
\arguments{
\item{dfLtDirect}{data frame, direct lifetime estimates}
}
\description{
Zero columns are required to be included in summations later
}
