% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyLtHs.R
\name{.prepareShare}
\alias{.prepareShare}
\title{Adjust the column names to obtain proper numbering of duplicates
and convert time period columns to factors}
\usage{
.prepareShare(df)
}
\arguments{
\item{df}{data frame for which the adjustment of column names and time periods should be done}
}
\description{
This is targeted at the shares read from gdx
}
