% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.prepVar}
\alias{.prepVar}
\title{Prepare variable}
\usage{
.prepVar(x, factor = 0.001, onlyArea = TRUE)
}
\arguments{
\item{x}{MAgPIE object of brick variable}

\item{factor}{numeric, scalar factor that \code{x} is multiplied with}

\item{onlyArea}{logical, if TRUE, select 'area' quantity and collapse
quantity dimension}
}
\value{
scales and filtered MAgPIE object
}
\description{
Convert values by multiplying with given factor (usually 1/1000 to get from
million m2 to billion m2) and select the area quantity.
}
\author{
Robin Hasse
}
