% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtMixed.R
\name{.processLtMixed}
\alias{.processLtMixed}
\title{Compute the relative value of the lifetime estimates}
\usage{
.processLtMixed(dfLt, dfVal, dfAnte, dims, removeVin = FALSE)
}
\arguments{
\item{dfLt}{data frame, mixed lifetime estimates}

\item{dfVal}{data frame, total inflows as given by brick results}

\item{dfAnte}{data frame, ex-ante lifetime estimates}

\item{dims}{character, dimensions of the data without time periods}

\item{removeVin}{logical, whether to remove the vintage column from the dimensions}
}
\description{
By setting the mixed estimates in relation to the total inflows, compute the
probabilities of removal in a given time period.
If both the total inflow and the absolute lifetime estimate are close to zero,
set this to the ex-ante value.
}
