% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtPost.R
\name{.processLtPost}
\alias{.processLtPost}
\title{Compute the relative value of the lifetime estimates}
\usage{
.processLtPost(ltPost, data, dims)
}
\arguments{
\item{ltPost}{data frame, ex-post lifetime estimates}

\item{data}{data frame, total inflows as given by brick results}

\item{dims}{character, dimensions of the data without time periods}
}
\description{
By setting the ex-post estimates in relation to the total inflows, compute the
probabilities of removal in a given time period.
If the total inflow is close to zero, this will be NA.
}
