% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAgg.R
\name{.readBrickSets}
\alias{.readBrickSets}
\title{Read brickSets mapping}
\usage{
.readBrickSets(tmpl)
}
\arguments{
\item{tmpl}{character, BRICK reporting template}
}
\value{
named list with definition of common set elements
}
\description{
Read brickSets mapping
}
