% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.readGdxIter}
\alias{.readGdxIter}
\title{Read a symbol from several gdx files and combine in one data frame}
\usage{
.readGdxIter(
  gdx,
  symbol,
  maxIter,
  minIter = 0,
  asMagpie = TRUE,
  ttotFilter = NULL,
  replaceVar = FALSE,
  dims = NULL
)
}
\arguments{
\item{gdx}{Path to gdx files; all gdx file names are assumed to be numbered and
adhere to the structure: If \code{gdx = <path/to/file/name.gdx}, the function looks for
\code{<path/to/file/name_0.gdx>}, \code{<path/to/file/name_1.gdx>}, ...}

\item{symbol}{Symbol to be read from the gdxes}

\item{maxIter}{Last iteration to be read}

\item{minIter}{First iteration to be read}

\item{asMagpie}{logical, convert to magpie object?}

\item{ttotFilter}{numeric/factor, time periods to filter for}

\item{replaceVar}{logical, replace column names \code{bs} and \code{hs} by \code{bsr} and \code{hsr}?}

\item{dims}{character, containing the colum names expected for the object that is read in.
Only required if the object does not exist to create replacement with NAs}
}
\value{
data frame of all results read in
}
\description{
Read a symbol from several gdx files and combine in one data frame
}
