% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyLtHs.R
\name{.removeColNumbering}
\alias{.removeColNumbering}
\title{Remove column numbering and renumber duplicates}
\usage{
.removeColNumbering(df, offset = 0, suffix = "")
}
\arguments{
\item{df}{data frame for which the columns should be renamed}

\item{offset}{numeric, number by which the renumbering is shifted}

\item{suffix}{character that is added the renumbered column name before the number}
}
\description{
Converts data with duplicates, for which all columns are numbered when reading from gdx,
to a state where only the duplicates are numbered.
}
