% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportHeatingSystemDetails.R
\name{.renameCostType}
\alias{.renameCostType}
\title{Rename the entries of the cost type column to more legible names}
\usage{
.renameCostType(df)
}
\arguments{
\item{df}{data frame with cost type column to be modified}
}
\description{
Rename the entries of the cost type column to more legible names
}
