% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.replaceVarName}
\alias{.replaceVarName}
\title{Replace column names \code{bs} and \code{hs} by \code{bsr} and \code{hsr}}
\usage{
.replaceVarName(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame where column names have been replaced
}
\description{
Replace column names \code{bs} and \code{hs} by \code{bsr} and \code{hsr}
}
