% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAgg.R
\name{.select}
\alias{.select}
\title{select values from MAgPIE-obect}
\usage{
.select(x, ...)
}
\arguments{
\item{x}{MAgPIE object}

\item{...}{entry selection. Combined dimensions have to be seperated with
\code{.} for both the set names and the set elements.}
}
\value{
MAgPIE object containing only selected entries
}
\description{
This is a wrapper around \code{magclass::mselect} that also allows the
selection of combinations of multiple dimensions.
}
