% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAgg.R
\name{.setNames}
\alias{.setNames}
\title{Wrapper around setNames}
\usage{
.setNames(object, nm)
}
\arguments{
\item{object}{MAgPIE object}

\item{nm}{a vector of names current names should be replaced with. If only
one data element exists you can also set the name to NULL.}
}
\value{
NULL if object is NULL else object with manipulated names
}
\description{
Wrapper around setNames
}
