% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.split}
\alias{.split}
\title{Split dimension names}
\usage{
.split(x, split = "\\\\.")
}
\arguments{
\item{x}{character vector}

\item{split}{character used to split \code{x}}
}
\value{
character vector with each dimension as an own entry
}
\description{
Split each entry of a character vector and return one unnested character
vector.
}
\author{
Robin Hasse
}
