% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLCCShare.R
\name{.splitDataByHs}
\alias{.splitDataByHs}
\title{Separate a large data set into several variables by hs}
\usage{
.splitDataByHs(data, hsNames, varName)
}
\arguments{
\item{data}{data frame to be split}

\item{varName}{character, name stem of resulting variable names}

\item{hsrNames}{character, hsr values to filter for}
}
\description{
Separate a large data set into several variables by hs
}
