% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendPeriods.R
\name{extendPeriods}
\alias{extendPeriods}
\title{Crop MagPIE object to given periods and fill missing periods}
\usage{
extendPeriods(x, periods, fill = NA)
}
\arguments{
\item{x}{MagPIE object}

\item{periods}{numeric vector with target periods}

\item{fill}{value to fill missing period values}
}
\value{
MagPIE object with given periods in temporal dimension
}
\description{
Crop MagPIE object to given periods and fill missing periods
}
\author{
Robin Hasse
}
