% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateQuantShares.R
\name{generateQuantShares}
\alias{generateQuantShares}
\title{Prototype / place holder for price sensitivity plot}
\usage{
generateQuantShares()
}
\description{
Early trials on plotting logarithmic quantity shares vs. relative lcc
}
\author{
Ricarda Rosemann
}
