% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heading.R
\name{heading}
\alias{heading}
\title{Print Rmd heading}
\usage{
heading(text, level = 1)
}
\arguments{
\item{text}{character, header text or sector tag}

\item{level}{numeric, level of heading (number of #s)}
}
\description{
Print Rmd heading
}
