% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeLCC.R
\name{normalizeLCC}
\alias{normalizeLCC}
\title{Normalize LCC}
\usage{
normalizeLCC(dfLcc, dfLt, dfDt, flow)
}
\arguments{
\item{dfLcc}{data frame, lifecycle costs}

\item{dfLt}{data frame, lifetime estimate}

\item{dfDt}{data frame, lengths of time periods}
}
\description{
Normalize the life cycle costs by rescaling to the average lifetime across
all heating systems
}
\author{
Ricarda Rosemann
}
