% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBRICKAdditional.R
\name{plotBRICKAdditional}
\alias{plotBRICKAdditional}
\title{Render additional Brick plotting routine}
\usage{
plotBRICKAdditional(
  path = ".",
  file = NULL,
  plottingRoutine = "plotsCalibration.Rmd",
  outName = "",
  scenNames = NULL,
  regionHandling = "all"
)
}
\arguments{
\item{path}{(named) character vector, path(s) to output directories.
If several paths are given, the names can be used to pass short scenario names.}

\item{file}{character vector, name(s) of file(s) with reporting results to be plotted.
- For \code{plotsCalibration.Rmd}, this is usually the file \code{BRICK_calibration_report.csv}
- For \code{plotsLcc.Rmd}, this is usually the file \code{BRICK_analysis_report.csv}}

\item{plottingRoutine}{Name of Rmd file to be rendered to generate the desired plots.
- \code{plotsCalibration.Rmd} generates plots of calibration results
- \code{plotsLcc.Rmd} generates analysis plots including life time assessments and LCC results}

\item{outName}{character, string added to the pdf file name}

\item{scenNames}{character vector, scenario names for different paths.
Needs to be specified if \code{path} is unnamed and contains more than one element.}

\item{regionHandling}{character, indication of how to handle the region dimensions. Available options:
- \code{all}: Take into account all regions for the report. With more than one region, this is
  only available for \code{plotsCalibration.Rmd}. If this option is chosen for \code{plotsLcc.Rmd}
  with more than one region, this parameter is reset to \code{"separate"}.
- \code{separate}: Produce a separate report for each region. This is the recommended setting
  to analyse all regions with \code{plotsLcc.Rmd}.
- \code{<character vector of regions>}: Only include the given regions in the report.
  For \code{plotsLcc.Rmd}, a separate report is generated for each region.}
}
\description{
Renders specified Rmd file to create a pdf of additional Brick plots.
The Rmd file needs to be present in \code{inst/plotsAdditional}.
Currently available files are:
  - plotsCalibration.Rmd
  - plotsLcc.Rmd
Mutiple scenarios and more than one region in one report are only supported for \code{plotsCalibration.Rmd}.
}
\author{
Ricarda Rosemann
}
