% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBRICKcalib.R
\name{plotBRICKCalib}
\alias{plotBRICKCalib}
\title{Render the BRICK calibration plotting routine}
\usage{
plotBRICKCalib(
  outputDir = NULL,
  cal = "BRICK_calibration_report.csv",
  outName = "",
  scenNames = NULL,
  savePlots = FALSE
)
}
\arguments{
\item{outputDir}{(named) character vector, path(s) to output directories.
If several paths are given, the names can be used to pass short scenario names.}

\item{cal}{character vector, name(s) of file(s) with calibration reporting results}

\item{outName}{character, string added to the pdf file name and names of additionally saved plots}

\item{scenNames}{character vector, scenario names for different paths.
Needs to be specified if \code{outputDir} is unnamed and contains more than one element.}

\item{savePlots}{logical, whether all plots should additionally be saved as png}
}
\description{
Renders the file plotCalibration.Rmd to create the plots for the BRICK calibration
}
\author{
Ricarda Rosemann
}
