% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatching.R
\name{plotMatching}
\alias{plotMatching}
\title{Matching reporting}
\usage{
plotMatching(path, outputFolder = NULL)
}
\arguments{
\item{path}{character, path to the run}

\item{outputFolder}{character, directory where output file is stored. If
NULL, the plots folder inside the given path is used.}
}
\value{
file path to created output file
}
\description{
Plot the following matching reporting
\itemize{
  \item comparison of the matching values and reference targets
  \item share of initial heating systems thar have been removed over time
}
}
\author{
Robin Hasse
}
