% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatchingComparison.R
\name{plotMatchingComparison}
\alias{plotMatchingComparison}
\title{Compare matching references with model values}
\usage{
plotMatchingComparison(path, outputFolder = NULL)
}
\arguments{
\item{path}{character, path to the run}

\item{outputFolder}{cahracter, directory where output file is stored. If
NULL, the plots folder inside the given path is used.}
}
\value{
file path to created output file
}
\description{
Compare matching references with model values
}
\author{
Robin Hasse
}
