% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBrickSets.R
\name{readBrickSets}
\alias{readBrickSets}
\title{Read brickSets mapping}
\usage{
readBrickSets(tmpl = NULL)
}
\arguments{
\item{tmpl}{character, BRICK reporting template}
}
\value{
named list with definition of common set elements
}
\description{
This function creates an explicit named list with the elements of all BRICK
dimensions and corresponding reporting names.
}
\author{
Robin Hasse
}
