% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBuildingStock.R
\name{reportBuildingStock}
\alias{reportBuildingStock}
\title{Report building Stock}
\usage{
reportBuildingStock(gdx, tmpl = NULL)
}
\arguments{
\item{gdx}{gams transfer container of the BRICK GDX}

\item{tmpl}{character, BRICK reporting template}
}
\description{
Report quantities describing the stock of buildings
}
\author{
Robin Hasse
}
