% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{reportCalibration}
\alias{reportCalibration}
\title{Read in model results from calibration for each iteration, calculate deviation}
\usage{
reportCalibration(gdx)
}
\arguments{
\item{gdx}{path to a gdx; it is assumed that for each iteration a gdx is present
with this path and the iteration number inserted at the end.}
}
\description{
Read in model results from calibration for each iteration, calculate deviation
}
\author{
Ricarda Rosemann
}
