% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportConstruction.R
\name{reportConstruction}
\alias{reportConstruction}
\title{Report construction}
\usage{
reportConstruction(gdx, brickSets = NULL, silent = TRUE)
}
\arguments{
\item{gdx}{gams transfer container of the BRICK GDX}

\item{brickSets}{character, BRICK reporting template}

\item{silent}{boolean, suppress warnings and printing of dimension mapping}
}
\description{
Report quantities describing the construction of new buildings
}
\author{
Robin Hasse
}
