% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDemandVars.R
\name{reportDemandVars}
\alias{reportDemandVars}
\title{Report energy demand and emissions}
\usage{
reportDemandVars(x, var, unit, brickSets, silent)
}
\arguments{
\item{x}{MagPIE object}

\item{var}{character, high-level variable name}

\item{unit}{character, unit to be pasted at the end of the variable name}

\item{brickSets}{character, BRICK reporting template}

\item{silent}{logical, suppress warnings and printing of dimension mapping}
}
\value{
MagPIE object with reporting data
}
\description{
Report energy demand and emissions
}
\author{
Robin Hasse
}
