% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDemolition.R
\name{reportDemolition}
\alias{reportDemolition}
\title{Report demolition}
\usage{
reportDemolition(gdx, brickSets = NULL, silent = TRUE)
}
\arguments{
\item{gdx}{gams transfer container of the BRICK GDX}

\item{brickSets}{character, BRICK reporting template}

\item{silent}{boolean, suppress warnings and printing of dimension mapping}
}
\description{
Report quantities describing the demolition of buildings
}
\author{
Robin Hasse
}
