% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEmissions.R
\name{reportEmissions}
\alias{reportEmissions}
\title{Report emissions}
\usage{
reportEmissions(gdx, brickSets = NULL, silent = TRUE)
}
\arguments{
\item{gdx}{gams transfer container of the BRICK GDX}

\item{brickSets}{character, BRICK reporting template}

\item{silent}{boolean, suppress warnings and printing of dimension mapping}
}
\description{
Report CO2 emissions from space heating
}
\author{
Robin Hasse
}
