% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportEnergy.R
\name{reportEnergy}
\alias{reportEnergy}
\title{Report energy demand}
\usage{
reportEnergy(gdx, brickSets = NULL)
}
\arguments{
\item{gdx}{gams transfer container of the BRICK GDX}

\item{brickSets}{character, BRICK reporting template}
}
\description{
Report final and useful energy demand for space heating
}
\author{
Robin Hasse
}
