% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportHeatingSystemDetails.R
\name{reportHeatingSystemDetails}
\alias{reportHeatingSystemDetails}
\title{Compute various heating system characteristics}
\usage{
reportHeatingSystemDetails(
  path,
  gdxName = "output.gdx",
  filterFullRen = list(vin = "1980-1989", ttotIn = 2015),
  hsRef = "gabo"
)
}
\arguments{
\item{path}{character, path to the Brick model output folder}

\item{gdxName}{character, file name of the gdx to read data from}

\item{filterFullRen}{named list with name value pairs to filter full resolution
renovation LCC/LCOH/Shares by.
If several filtering criteria are given, the data is filtered to always match at least one criterion.
If \code{NULL}, no filtering is applied.}

\item{hsRef}{character, reference hs for linearized logistic model}
}
\description{
Estimate heating system lifetimes and use this to compute lifecycle costs,
levelized costs of heat and logit heating system shares.
Extract brick model shares for comparison.
}
\author{
Ricarda Rosemann
}
