% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPriceInput.R
\name{reportPriceInput}
\alias{reportPriceInput}
\title{Read in input price data and heat pump share}
\usage{
reportPriceInput(
  path,
  gdxName = "input.gdx",
  gdxNameReferences = "references.gdx"
)
}
\arguments{
\item{path}{character, path to the desired input gdx}

\item{gdxName}{character, file name of the desired input gdx}
}
\description{
Read in input price data and heat pump share
}
