% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportRates.R
\name{reportRates}
\alias{reportRates}
\title{Report renovation and exchange rates}
\usage{
reportRates(gdx, brickSets = NULL, silent = TRUE)
}
\arguments{
\item{gdx}{gams transfer container of the BRICK GDX}

\item{brickSets}{character, BRICK reporting template}

\item{silent}{boolean, suppress warnings and printing of dimension mapping}
}
\description{
Yearly renovation activity relative to the stock
}
\author{
Robin Hasse
}
