% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportRenovationBS.R
\name{reportRenovationBS}
\alias{reportRenovationBS}
\title{Report renovation of the building shell}
\usage{
reportRenovationBS(gdx, brickSets = NULL, silent = TRUE)
}
\arguments{
\item{gdx}{gams transfer container of the BRICK GDX}

\item{brickSets}{character, BRICK reporting template}

\item{silent}{boolean, suppress warnings and printing of dimension mapping}
}
\description{
Report quantities describing the renovation of the building shell
}
\author{
Ricarda Rosemann
}
