% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportRenovationHS.R
\name{reportRenovationHS}
\alias{reportRenovationHS}
\title{Report renovation of heating systems}
\usage{
reportRenovationHS(gdx, brickSets = NULL, silent = TRUE)
}
\arguments{
\item{gdx}{gams transfer container of the BRICK GDX}

\item{brickSets}{character, BRICK reporting template}

\item{silent}{boolean, suppress warnings and printing of dimension mapping}
}
\description{
Report quantities describing the renovation of heating systems
}
\author{
Ricarda Rosemann
}
