% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleRemainingShare.R
\name{rescaleRemainingShare}
\alias{rescaleRemainingShare}
\title{Rescale the remaining share to a given Weibull inflow}
\usage{
rescaleRemainingShare(remainingShare, wbRescale, ttotRescale, dims)
}
\arguments{
\item{remainingShare}{data frame with share of remaining installations}

\item{wbRescale}{data frame with rescaling factor from weibull distribution}

\item{ttotRescale}{numeric, time period to rescale on}
}
\description{
Rescale the remaining share to a given Weibull inflow
}
\author{
Ricarda Rosemann
}
