% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showAnalysisPlot.R
\name{showAnalysisPlot}
\alias{showAnalysisPlot}
\title{Plotting function, mainly for LCC analysis plots}
\usage{
showAnalysisPlot(
  plotType,
  data,
  varName,
  yname,
  color = NULL,
  linetype = NULL,
  facets = c("loc", "typ"),
  rprt = NULL,
  avg = NULL,
  remCols = NULL,
  xname = "ttotOut",
  valueName = yname,
  suppressLateTtot = TRUE,
  xlabName = NULL,
  ylabName = NULL,
  tmpl = NULL,
  addPoints = NULL,
  filterRows = list(hs = "h2bo", hsr = "h2bo"),
  valueCap = Inf
)
}
\arguments{
\item{plotType}{character, desired plot type.
Needs to be one of \code{line}, \code{bar} or \code{twoBar}.}

\item{data}{data frame containing the data to be plotted}

\item{varName}{character, variables to be plotted}

\item{yname}{character, column to be used as y-aesthetic}

\item{color}{character, column to be used as color aesthetic}

\item{linetype}{character, column to be used as linetype aesthetic}

\item{facets}{character, columns to be used as facets.
Can either be one or two columns.}

\item{rprt}{character, columns for which a separate plot should be created for
each entry}

\item{avg}{character, columns over which the average should be computed}

\item{remCols}{character, columns to be removed before creating the plot}

\item{xname}{character, column to be used as x-aesthetic}

\item{suppressLateTtot}{logical, whether to remove the last three time steps
from the plots shown}

\item{xlabName}{character, x-axis label}

\item{ylabName}{character, y-axis label}

\item{tmpl}{character, name or path to reportbrick reporting template.}
}
\description{
Framework to create various plots required for the LCC analysis.
Available plot types are:
 - line plot
 - (stacked) bar plot
 - comparison bar plot with two stacked bars of different transparency
}
\author{
Ricarda Rosemann
}
