% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMatchingComparison.R
\name{showMatchingComparison}
\alias{showMatchingComparison}
\title{Plot comparison between matching references and model values}
\usage{
showMatchingComparison(path, showTitles = TRUE)
}
\arguments{
\item{path}{character, path to the run}

\item{showTitles}{logical, if TRUE, the references name and basic variable
name are shown as title in each plot. No titles otherwise.}
}
\value{
named list of ggplot2 objects
}
\description{
Create stacked bar plots to verify matching accuracy
}
\author{
Robin Hasse
}
