% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showStockAndFlows.R
\name{showStockAndFlows}
\alias{showStockAndFlows}
\title{Stacked bar plot summarising the stock evolution}
\usage{
showStockAndFlows(
  data,
  brickSets = NULL,
  vars = c("Stock", "Construction", "Demolition"),
  cols = NULL,
  fill = "hs",
  identTransp = FALSE,
  revStackOrder = FALSE
)
}
\arguments{
\item{data}{data.frame from BRICK mif}

\item{brickSets}{character, BRICK reporting template}

\item{vars}{variables shown as rows}

\item{cols}{subsets of the building stock shown as columns}

\item{fill}{characteristi mapped to the fill colour}

\item{identTransp}{logical, if TRUE, identical replacement in renovation is
rendered semi transparent}

\item{revStackOrder}{boolean, reverse stacking order?}
}
\value{
ggplot2 object with bar plot
}
\description{
Stacked bar plot summarising the stock evolution
}
\author{
Robin Hasse
}
