% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyLtHs.R
\name{verifyLtHs}
\alias{verifyLtHs}
\title{Verify the lifetime constraint of heating systems}
\usage{
verifyLtHs(inflow, v_stockInit, outflow, p_shareRenHSinit, p_shareRenHS, dims)
}
\arguments{
\item{inflow}{data frame, all inflows}

\item{v_stockInit}{data frame, brick results on initial stock}

\item{outflow}{data frame, all outflows}

\item{p_shareRenHSinit}{data frame with shares to be renovated for the initial stock}

\item{p_shareRenHS}{data frame with shares to be renovated for the inflow}

\item{dims}{character, dimensions of the data without time periods}
}
\description{
Verify the lifetime constraint of heating systems
}
\author{
Ricarda Rosemann
}
