% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMatching.R
\name{checkMatching}
\alias{checkMatching}
\title{Check the matching results for inconsistencies}
\usage{
checkMatching(path, tol = 0.01)
}
\arguments{
\item{path}{character, path to the matching result files}

\item{tol}{numeric, tolerance against which to check the stock balance equations}
}
\description{
Check the matching results for inconsistencies
}
\author{
Ricarda Rosemann
}
