% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtAnte.R
\name{computeLtAnte}
\alias{computeLtAnte}
\title{Compute heating system lifetimes ex-ante}
\usage{
computeLtAnte(data, shareRen, t0, isFlow = TRUE)
}
\arguments{
\item{data}{data frame, dimensions of desired lifetimes,
usually also contains values to compute the absolute removed quantities}

\item{shareRen}{data frame with the data on which share of inflows has to be replaced when}

\item{t0}{numeric, initial time period (time period before the model starts)}

\item{isFlow}{logical, is \code{data} flow data?}
}
\description{
Determine the lifetime of the heating system according to the respective Weibull distribution.
No early retirement is reflected
}
\author{
Ricarda Rosemann
}
