% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtAnte.R
\name{computeLtAnte}
\alias{computeLtAnte}
\title{Compute heating system lifetimes ex-ante}
\usage{
computeLtAnte(
  variable,
  data,
  ttotNum,
  lifeTimeHs,
  dims,
  runSimple = FALSE,
  dataValue = NULL,
  p_dt = NULL
)
}
\arguments{
\item{variable}{character, variable for which the lifetime is computed}

\item{data}{data frame, dimensions of desired lifetimes,
usually also contains values to compute the absolute removed quantities}

\item{ttotNum}{numeric, all time periods to be considered}

\item{lifeTimeHs}{data frame, Weibull parameters}

\item{dims}{character, dimensions of the data without time dimensions}

\item{runSimple}{logical, whether to use the simplified formula}

\item{dataValue}{data frame, optionally provide the values to compute the
absolute removed quantities separately}

\item{p_dt}{data frame, lengths of the time periods
(only required for simple formula)}
}
\description{
Determine the lifetime of the heating system according to the respective Weibull distribution.
No early retirement is reflected
}
\author{
Ricarda Rosemann
}
