% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRemainingShare.R
\name{computeRemainingShare}
\alias{computeRemainingShare}
\title{Compute the remaining share of an inflow}
\usage{
computeRemainingShare(dfLt, asDensity = TRUE, valueName = "relVal")
}
\arguments{
\item{dfLt}{data frame with life time data}

\item{asDensity}{logical, whether the data is given as a density}

\item{valueName}{character, name of the column from which the renovation share is read}
}
\description{
Compute the remaining share of an inflow
}
\author{
Ricarda Rosemann
}
