% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLCCShare.R
\name{.checkLifetimeResults}
\alias{.checkLifetimeResults}
\title{Check plausibility of lifetime results}
\usage{
.checkLifetimeResults(dfLt, dims)
}
\description{
Check plausibility of lifetime results
}
