% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLCCShare.R
\name{.cleanReadGdx}
\alias{.cleanReadGdx}
\title{Read data from gdx and clean}
\usage{
.cleanReadGdx(
  gdx,
  symbol,
  vinExists = NULL,
  renAllowed = NULL,
  asMagpie = FALSE
)
}
\arguments{
\item{gdx}{path to the gdx}

\item{symbol}{character, symbol to read from gdx}

\item{vinExists}{data frame with existing vintage and time period combinations}

\item{renAllowed}{data frame with all possible renovation transitions}

\item{asMagpie}{logical, whether to convert the data to a magpie opject}
}
\description{
Read data from gdx and clean
}
