% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLCCShare.R
\name{.cleanReadGdx}
\alias{.cleanReadGdx}
\title{Read in Gdx and slightly modify}
\usage{
.cleanReadGdx(df, vinExists, renAllowed = NULL, bsToZero = TRUE)
}
\description{
Read in Gdx and slightly modify
}
