% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtMixed.R
\name{.combineAddAttr}
\alias{.combineAddAttr}
\title{Add the additional attributions of two data frames}
\usage{
.combineAddAttr(dfThis, dfOther, dims)
}
\arguments{
\item{dfThis}{data frame, contains a column \code{addAttr} with the additional attributions}

\item{dfOther}{data frame, contains a column \code{addAttr} with the additional attributions}

\item{dims}{character, dimensions of the data without the time periods}
}
\description{
This is required to join the additional attributions of construction and
renovation flows.
}
