% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.computeAvg}
\alias{.computeAvg}
\title{Compute the mean value in a data frame}
\usage{
.computeAvg(df, rprt = "")
}
\arguments{
\item{df}{data frame, containing the data to be evaluated}

\item{rprt}{character, column names for which the mean should be reported separately}
}
\value{
data frame averages as value column
}
\description{
Compute the mean value in a data frame
}
