% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.computeDeviation}
\alias{.computeDeviation}
\title{Compute the deviation to historical data}
\usage{
.computeDeviation(df, dfHist)
}
\arguments{
\item{df}{data frame, containing data from calibration run}

\item{dfHist}{data frame, containing historical data}
}
\value{
data frame where the value column contains the deviation from historical data
}
\description{
Compute the deviation to historical data
}
