% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.computeFlowSum}
\alias{.computeFlowSum}
\title{Compute the sum of construction and renovation flow values}
\usage{
.computeFlowSum(con, ren)
}
\arguments{
\item{con}{data frame, contains construction flow quantities}

\item{ren}{data frame, contains renovation flow quantities}
}
\value{
data frame with the sum in the value column
}
\description{
Compute the sum of construction and renovation flow values
}
