% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLCC.R
\name{.computeLCCOpe}
\alias{.computeLCCOpe}
\title{Compute lifecycle operational costs}
\usage{
.computeLCCOpe(dfLt, dfCosts, dfDt, dfDiscount)
}
\arguments{
\item{dfLt}{data frame, lifetime estimate}

\item{dfCosts}{data frame, costs}

\item{dfDiscount}{data frame, discount factor}
}
\description{
Compute lifecycle operational costs
}
