% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtMixed.R
\name{.computeLtDirect}
\alias{.computeLtDirect}
\title{Compute direct lifetime estimate}
\usage{
.computeLtDirect(dfLtAnte, outAnte, outflow, dims)
}
\arguments{
\item{dfLtAnte}{data frame, ex-ante lifetimes}

\item{outAnte}{data frame, total ex-ante outflows}

\item{outflow}{data frame, total actual brick outflows}

\item{dims}{character, dimensions of the data without time periods}
}
\description{
Compute the lifetime as the ex-ante estimate scaled by the ratio of ex-ante
and actual total outflows
}
