% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.computeSum}
\alias{.computeSum}
\title{Compute the sum}
\usage{
.computeSum(df, rprt = "")
}
\arguments{
\item{df}{data frame containing the data to be evaluated}

\item{rprt}{character, column names for which the sum should be reported separately}
}
\value{
data frame with summed values
}
\description{
Compute the sum
}
