% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.computeSumSq}
\alias{.computeSumSq}
\title{Compute the sum of the squares in a data frame}
\usage{
.computeSumSq(df, rprt = "")
}
\arguments{
\item{df}{data frame, containing the data to be evaluated}

\item{rprt}{character, column names for which the sum of the squares should be reported separately}
}
\value{
data frame sum of squares as value column
}
\description{
Compute the sum of the squares in a data frame
}
