% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLtMixed.R
\name{.computeUnattrOut}
\alias{.computeUnattrOut}
\title{Compute the unattributed outflow}
\usage{
.computeUnattrOut(dfLtMixed, dfTotVal, tOut, dims)
}
\arguments{
\item{dfLtMixed}{data frame, direct lifetime estimates of current time period and
mixed lifetime estimates of previous time periods}

\item{dfTotVal}{data frame, inflows in brick results}

\item{tOut}{numeric, time period of removal for which we are computing the
unattributed outflows}

\item{dims}{character, dimensions of the data without time periods}
}
\description{
Compute the amount of the brick outflow that has not been attributed to an
inflow by the direct lifetime estimate when also reflecting what remains of
the inflow at the current time period.
}
