% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCalibration.R
\name{.getDims}
\alias{.getDims}
\title{Get dimension names from stock and flow objects}
\usage{
.getDims(calibObj, removeDims = NULL)
}
\arguments{
\item{calibObj}{list, calibration targets}

\item{removeDims}{character, additional dimensions to be removed from the output}
}
\value{
character with dimension names
}
\description{
Gives the column names of a list of data frames except the first and the last
one, i.e. \code{qty} and \code{value}.
}
