% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizePriceSensitivity.R
\name{normalizePriceSensitivity}
\alias{normalizePriceSensitivity}
\title{Compute a normalized price sensitivity from life-cycle costs}
\usage{
normalizePriceSensitivity(dfLcc, brickRes, lambda, dims, groupCols = NULL)
}
\description{
Compute a normalized price sensitivity from life-cycle costs
}
\author{
Ricarda Rosemann
}
