% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBRICKAdditional.R
\name{plotBRICKAdditional}
\alias{plotBRICKAdditional}
\title{Render additional Brick plotting routine}
\usage{
plotBRICKAdditional(
  path = ".",
  file = "BRICK_calibration_report.csv",
  plottingRoutine = "plotsCalibration.Rmd",
  outName = "",
  scenNames = NULL,
  savePlots = FALSE
)
}
\arguments{
\item{path}{(named) character vector, path(s) to output directories.
If several paths are given, the names can be used to pass short scenario names.}

\item{file}{character vector, name(s) of file(s) with reporting results to be plotted.
- For \code{plotsCalibration.Rmd}, this is usually the file \code{BRICK_calibration_report.csv}
- For \code{plotsLcc.Rmd}, this is usually the file \code{BRICK_analysis_report.csv}}

\item{outName}{character, string added to the pdf file name and names of additionally saved plots}

\item{scenNames}{character vector, scenario names for different paths.
Needs to be specified if \code{path} is unnamed and contains more than one element.}

\item{savePlots}{logical, whether all plots should additionally be saved as png}
}
\description{
Renders specified Rmd file to create a pdf of additional Brick plots.
The Rmd file needs to be present in \code{inst/plotsAdditional}. plotCalibration.Rmd to create the plots for the BRICK calibration
Currently available files are:
  - plotsCalibration.Rmd
  - plotsLcc.Rmd
Mutiple scenarios are only supported for \code{plotsCalibration.Rmd}.
}
\author{
Ricarda Rosemann
}
