% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGdxSymbol.R
\name{readGdxSymbol}
\alias{readGdxSymbol}
\title{Read symbol from gams container}
\usage{
readGdxSymbol(
  gdx,
  symbol,
  field = "level",
  asMagpie = TRUE,
  stringAsFactor = TRUE
)
}
\arguments{
\item{gdx}{character, file path to GDX file}

\item{symbol}{character, name of gams object}

\item{field}{character, field to read (only relevant for variables)}

\item{asMagpie}{boolean, return Magpie object}

\item{stringAsFactor}{logical, keep default factors from gams}
}
\value{
MagPIE object with data of symbol
}
\description{
Read symbol from gams container
}
\author{
Robin Hasse
}
