% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAgg.R
\name{reportAgg}
\alias{reportAgg}
\title{Report aggregated quantities}
\usage{
reportAgg(
  x,
  name,
  brickSets = readBrickSets(NULL),
  agg = NULL,
  rprt = NULL,
  silent = TRUE
)
}
\arguments{
\item{x}{MagPIE object, BRICK object}

\item{name}{character, name of reporting variable. reported dimensions passed
with \code{rprt} have to be escaped with curly brackets.}

\item{brickSets}{named list, BRICK reporting template}

\item{agg}{named vector of dimensions to aggregate. Names are dimension names
of \code{x} and values are either set elements or subsets of set elements
to consider.}

\item{rprt}{named vector of dimensions to report individually. Names are
dimension names of \code{x} and values are either set elements or subsets
of set elements to report.}

\item{silent}{boolean, suppress warnings and printing of dimension mapping}
}
\description{
Report aggregated quantities
}
\note{
To consider specific combinations of dimensions in both \code{agg} or
  \code{rprt}, the combination can be given just like one dimension
  separated by \code{.}.
}
\author{
Robin Hasse
}
