% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBrick-package.R
\docType{package}
\name{reportBrick-package}
\alias{reportBrick}
\alias{reportBrick-package}
\title{reportBrick: Reporting package for BRICK}
\description{
This package contains BRICK-specific routines to report model results. The main functionality is to generate a mif-file from a given BRICK model run folder.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/reportBrick}
}

}
\author{
\strong{Maintainer}: Robin Hasse \email{robin.hasse@pik-potsdam.de} (\href{https://orcid.org/0000-0003-1818-3186}{ORCID})

}
\keyword{internal}
