% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportbrick-package.R
\docType{package}
\name{reportbrick-package}
\alias{reportbrick}
\alias{reportbrick-package}
\title{reportbrick: Reporting package for BRICK}
\description{
This package contains BRICK-specific routines to report model results. The main functionality is to generate a mif-file from a given BRICK model run folder.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/reportbrick}
}

}
\author{
\strong{Maintainer}: Robin Hasse \email{robin.hasse@pik-potsdam.de} (\href{https://orcid.org/0000-0003-1818-3186}{ORCID})

Authors:
\itemize{
  \item Ricarda Rosemann \email{ricarda.rosemann@pik-potsdam.de} (\href{https://orcid.org/0009-0006-5939-3197}{ORCID})
}

}
\keyword{internal}
