% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBuildingStock.R
\name{reportBuildingStock}
\alias{reportBuildingStock}
\title{Report building Stock}
\usage{
reportBuildingStock(gdx, brickSets = NULL, silent = TRUE)
}
\arguments{
\item{gdx}{gams transfer container of the BRICK GDX}

\item{brickSets}{character, BRICK reporting template}

\item{silent}{boolean, suppress warnings and printing of dimension mapping}
}
\description{
Report quantities describing the stock of buildings
}
\author{
Robin Hasse
}
