% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportConstruction.R
\name{reportConstruction}
\alias{reportConstruction}
\title{Report construction}
\usage{
reportConstruction(gdx, tmpl = NULL)
}
\arguments{
\item{gdx}{gams transfer container of the BRICK GDX}

\item{tmpl}{character, BRICK reporting template}
}
\description{
Report quantities describing the construction of new buildings
}
\author{
Robin Hasse
}
