% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDemolition.R
\name{reportDemolition}
\alias{reportDemolition}
\title{Report demolition}
\usage{
reportDemolition(gdx, tmpl = NULL)
}
\arguments{
\item{gdx}{gams transfer container of the BRICK GDX}

\item{tmpl}{character, BRICK reporting template}
}
\description{
Report quantities describing the demolition of buildings
}
\author{
Robin Hasse
}
