% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLCCShare.R
\name{reportLCCShare}
\alias{reportLCCShare}
\title{Compute life cycle costs and logit share}
\usage{
reportLCCShare(gdx, pathLt = NULL)
}
\description{
Compute life cycle costs and logit share
}
\author{
Ricarda Rosemann
}
