% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showAreaBarLinePlots.R
\name{showAreaBarLinePlots}
\alias{showAreaBarLinePlots}
\title{Show Area and bar plots followed by multiple line plots}
\usage{
showAreaBarLinePlots(
  data,
  tot,
  items,
  heading = NULL,
  nHeading = NULL,
  showLines = TRUE,
  showTot = TRUE
)
}
\arguments{
\item{data}{A quitte object}

\item{tot}{character, a total value to be shown in the area plots. If
multiple totals are passed, plots are created for each total.}

\item{items}{A character vector. Appending these items tot the total yields
the variables to plot.}

\item{heading}{character, header text}

\item{nHeading}{numeric, level of heading}

\item{showLines}{logical, should total line be plotted on are and bar plots?}
}
\description{
Show Area and bar plots followed by multiple line plots
}
\author{
Robin Hasse
}
