% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMatchingStandingStock.R
\name{showMatchingStandingStock}
\alias{showMatchingStandingStock}
\title{Plot share of initial heating systems removed}
\usage{
showMatchingStandingStock(path)
}
\arguments{
\item{path}{character, path to the run}
}
\value{
named list of ggplot2 objects
}
\description{
Show the share of heating systems standing in the initial time step that have
been removed over time. The tunnel marks the allowed space between the lower
and upper limit defined in \code{p_shareRenHSinit} and the solid line shows
the value eventually chosen in the matching by \code{v_shareRenHSinit}. The
dashed line shows the extrapolation of this result to future time steps that
go beyond the temporal scope of the matching. The dotted line marks the
central value that results from an evenly distributed installation up until
the initial time steps. The central value is not considered in the
optimisation but used as supporting point in the extrapolation.
}
\author{
Robin Hasse
}
