% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSankey.R
\name{showSankey}
\alias{showSankey}
\title{Plot Sankey diagram}
\usage{
showSankey(
  path,
  fill = c("bs", "hs"),
  showOnlyRenovation = FALSE,
  filterData = NULL,
  maxPeriodsInRow = NULL,
  save = TRUE
)
}
\arguments{
\item{path}{character, path to the run}

\item{fill}{character, dimension mapped to fill, either \code{"bs"} (building
shell) or \code{"hs"} (heating system).}

\item{showOnlyRenovation}{logical, if \code{FALSE}, the stock and all flows
are shown including untouched buildings. If \code{TRUE}, only renovation
activities are shown. This is helpful to better observe the renovation
decision, especially with short time steps where renovation only affects a
small share of the stock.}

\item{filterData}{named list to filter the data before plotting}

\item{maxPeriodsInRow}{maximum number of time steps to plot in one row. If
there is more data, it is broken across multiple lines. If \code{NULL}, all
data is shown in one line.}

\item{save}{control how the plot should be saved:
\itemize{
  \item `TRUE`: plot is saved in plots folder of run path
  \item `FALSE`: plot is not saved (but always returned invisibly)
  \item character: plot is saved in given directory
}}
}
\description{
Visualise transitions from one state to another. This plot helps to follow
the exact renovation flows, though it quickly gets very complex.
}
\author{
Robin Hasse
}
