% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyLtHs.R
\name{verifyLtHs}
\alias{verifyLtHs}
\title{Verify the lifetime constraint of heating systems}
\usage{
verifyLtHs(
  gdx,
  inflow,
  v_stockInit,
  outflow,
  lifeTimeHs,
  dims,
  standingLtStock = 6
)
}
\arguments{
\item{gdx}{character, path to input gdx file}

\item{inflow}{data frame, all inflows}

\item{v_stockInit}{data frame, brick results on initial stock}

\item{outflow}{data frame, all outflows}

\item{dims}{character, dimensions of the data without time periods}

\item{standingLtStock}{numeric, assumed prior standing lifetime of the initial stock}
}
\description{
Verify the lifetime constraint of heating systems
}
\author{
Ricarda Rosemann
}
